Dev trainer refactor

+++ ssljax/augment/augmentation/colortransform.py
@@ -1,6 +1,7 @@
-# functions for ssljax.augment.augmentation.ColorTransform
 import jax
 import jax.numpy as jnp
+
+# these implementations adapted from https://github.com/deepmind/deepmind-research/blob/fba48d1e44d86628b65a31549560b7be2a25d823/byol/utils/augmentations.py
 
 def rgb_to_hsv(r, g, b):
     """Converts R, G, B  values to H, S, V values.
@@ -49,7 +50,7 @@ def hsv_to_rgb(h, s, v):
     dh = (h % 1.0) * 6.0
     fmodu = dh % 2.0
     x = c * (1 - jnp.abs(fmodu - 1))
-    hcat = jnp.floor(dh).astype(jnp.int32)
+    hcat = jnp.floor(dh).astype(jnp.float32)
     rr = (
         jnp.where(
             (hcat == 0) | (hcat == 5), c, jnp.where((hcat == 1) | (hcat == 4), x, 0)
